+++ packages/graphico/example/test.mjs
@@ -1,2 +1,0 @@
-import { PKG_NAME } from 'graphico';
-console.log(PKG_NAME);