+++ packages/graphico/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.0.2
+
+- Add first iteration of `Canvas` class
+    - Support several callback functions (`keydown`/`up`, `mousedown`/`up`, `mousemove`)
+- Create a basic painting example
+
 ## 0.0.1
 
 - Add details in `package.json`